formatted

+++ src/image_io.py
@@ -1,5 +1,4 @@
 import numpy as np
-import cv2
 import PIL
 from PIL import Image
 import os
@@ -7,7 +6,9 @@ import glob
 from typing import Union, Tuple, Iterator
 
 
-def read_image(path: str, resize_ratio: Union[float, None] = None) -> Union[Tuple[None, None], Tuple[np.ndarray, Tuple[int, int]]]:
+def read_image(
+    path: str, resize_ratio: Union[float, None] = None
+) -> Union[Tuple[None, None], Tuple[np.ndarray, Tuple[int, int]]]:
     """
     Read image by gray scale.
 
@@ -23,14 +24,16 @@ def read_image(path: str, resize_ratio: Union[float, None] = None) -> Union[Tupl
     (None, None) or (img, dpi) : None or (np.ndarray, (int, int))
     """
     try:
-        pilimg = Image.open(path).convert("L") # read as gray scale
+        pilimg = Image.open(path).convert("L")  # read as gray scale
     except PIL.UnidentifiedImageError:
         return None, None
     dpi = pilimg.info["dpi"]
     if resize_ratio is not None:
-        pilimg = pilimg.resize((int(pilimg.width * resize_ratio), int(pilimg.height * resize_ratio)))
+        pilimg = pilimg.resize(
+            (int(pilimg.width * resize_ratio), int(pilimg.height * resize_ratio))
+        )
         dpi = (int(dpi[0] * resize_ratio), int(dpi[1] * resize_ratio))
-    return np.array(pilimg, dtype=np.uint8) , dpi
+    return np.array(pilimg), dpi
 
 
 def read_images(
@@ -65,7 +68,8 @@ def read_images(
             continue
         yield p, img, dpi
 
-def save_image(path: str, img : np.ndarray, dpi: Tuple[int, int]):
+
+def save_image(path: str, img: np.ndarray, dpi: Tuple[int, int]):
 
     pilimg = Image.fromarray(img)
     pilimg.save(path, dpi=dpi)
